library: outline VaList into ffi::va_list
and reexport
